added configuration for vertical/horizontal depth slider
Now the setting dialog box contains a checkbox to select if the
traditional horizontal depth slider shall be used or the new proposed
vertical one.
I decided that the change will take place only after a Minutor restart,
as it would be complicated to change the layout after setup.
The setting dialog is also converted to a Qt Form, which makes it easier
to add changes.