add FlatteningConverter
This is the first step to automatically generate a converter from
vanilla_ids.json
Module is reading ID and data from the definition file and generades a
static  palette that might be used by a backwards compatible chunk
reader. This artificial palette contains all 1.12 blocks with variants.
It is huge (4096 entries), but will be used only once.
Mapping names are guessed automatically for the moment. In case this
fails for some blocks, it is possible to give a hint without breaking
old Minutor releases (new tag "flatname").