migration from STRING to HASH to handle Blocks internally
Accessing a QMap and searching Blocks is very very slow because of the QString handling. Compared to the old code before 1.13 it has a penalty of 2x .. 3x rendering time for a map.

Using qHash to convert all QStrings into numbers and using this hached ID (hid) brings back the speed. I measured also 2x .. 3x performance gain and old maps render as fast as before.