Store zoom level as a double
This changes the way that the current zoom level is stored by the
program. Previously, only the integer index of the zoom level was
stored. This leads to an issue: when you have many scroll events
firing every second, each individual scroll event will only have
a tiny delta. Most of these deltas will not be large enough to
change the zoom level, but they should build up over time. The
previous implementation lost these deltas because of rounding.

With this change, we store the zoom level as a double, and only
convert this to an index (by rounding) when this is needed to
change the scaling factor.

I also fixed a couple of tiny issues:

1. changed delta -> angleDelta to fix QT deprecation warnings.

2. preventing having to call redraw() for every scroll event

Fixes #224