When zooming with pointing device, zoom into cursor location (#295)
Before this commit, zooming always zooms into the center of the
view. After this commit, we zoom in such a way that the point under
the cursor represents the same location on the map before and after
the zoom. The imitates the behavior of most online mapping tools,
e.g. OpenStreetMaps.

The new zooming approach only applies when using the scroll wheel,
not when zooming with the keyboard. It is also dependent on a
QSetting. If desired, a future commit could expose this setting in
the GUI, making it possible to disable the new zoom behavior.