performance optimization: first steps (#307)
* don't do multiple searches in BlockIdentifier::getBlockInfo()

BlockIdentifier::getBlockInfo() is one of (or even the most)
frequently called functions during map rendering, so it should be
fast as possible. but now it is not - it does search on map twice:
one for contains() call, and another one on operator[]. this can be
easily avoided, and this is obvious optimization.

* replace QMap with QHash in BlockIdentifier

this is further optimization of BlockIdentifier::getBlockInfo().
there is nothing that requires sorting / specific order, so having
map is not really necessary, so hash table will suit better, and
it has constant time for element access (what is crucial in there).