synchronize QAction check state with QCheckBox state
for QWidgetAction 'checkable' prperty is also can be set and it is
maintained by QAction logic (i.e. is set to true/false on click).

this allows to handle clicks on whatever part of menu item, not
only on 'embedded' QCheckBox itself. fortunatelly, we don't need
to handle 'interactive' (i.e. as result of click) partial state,
it is only set programmatically.