drop submenues on world reload
it is assumed that menues with all found structures are cleared on
world reload, but submenues created for modded items are kept.
let's remove them too.

it is enough just to destroy QMenu instances to remove (and delete)
corresponding QAction instances. interesting, in case of submenu,
submenu QMenu object owns QAction, not QAction owns that QMenu...