Keep zoom level and index in sync when doing memory snapping
When zooming out, sometimes the user will scroll out so far that
memory limits will force the zoom back in a level. Currently,
however, this only changes the zoomIndex variable which controls
the rendering. The "true" zoom stores fractional scroll values
(e.g. from touchpads), but this was left unchanged. The result
is that this stored zoom value can disagree with the apparent
zoom level shown on the screen.

Make the memory-based zoom snapping keep zoomLevel and zoomIndex
in sync.