Fix snapping zoom to minimum when scrolling (#356)
* Fix snapping zoom to minimum when scrolling

This code is designed to prevent zooming past the minimum or
maximum, but it doesn't work properly in one important case. When
zoom out (zoom levels < 1:1) is enabled via a modifier key, and
zoom out has occured (the current zoom is less than 0), and then
the modifier key is released, further attempts to zoom out will
instead zoom in as the zoom is snapped to the reduced maximum.

This is observable if you scroll by mistake, but the bigger issue
is with high resolution touchpads, when generate numerous scroll
events and usually have "momentum", meaning they continue
generating scroll events after the user's fingers are off the
touchpad. This means the user can do everything "right", and release
the modifier key after they stop scrolling, but the momentum scroll
events will result in an annoying snap-in zoom.

This commit works around the issue without side effects (I think):

 * We only snap the zoom level *in* when the user is zooming *out*.

 * In case the previous zoom level is lower than the current minimum
   zoom level, we keep the previous zoom level instead of snapping.

Also fixes a typo.

* Keep zoom level and index in sync when doing memory snapping

When zooming out, sometimes the user will scroll out so far that
memory limits will force the zoom back in a level. Currently,
however, this only changes the zoomIndex variable which controls
the rendering. The "true" zoom stores fractional scroll values
(e.g. from touchpads), but this was left unchanged. The result
is that this stored zoom value can disagree with the apparent
zoom level shown on the screen.

Make the memory-based zoom snapping keep zoomLevel and zoomIndex
in sync.